+++ src/routes/league_player.rs
@@ -20,10 +20,10 @@ pub async fn change_league_request_status(conn: Data<Arc<MySqlPool>>, client: Da
     league_player::change_league_request_status(&conn, &client, new_status.to_owned(), join_req.0).await
 }
 #[post("/leagues/{page}")]
-pub async fn get_all_leagues_player_has_applied_to(conn: Data<Arc<MySqlPool>>, client: Data<Arc<Client>>, join_req: Json<JoinRequest>, page: Path<i32>) -> TypedHttpResponse<Vec<League>> {
+pub async fn get_all_leagues_player_has_applied_to(conn: Data<Arc<MySqlPool>>, client: Data<Arc<Client>>, join_req: Json<JoinRequest>, page: Path<u16>) -> TypedHttpResponse<Vec<League>> {
     league_player::get_all_leagues_player_has_applied_to(&conn, &client, join_req.0, *page).await
 }
-#[post("/players")] //TODO: Add page to this
+#[post("/players")]
 pub async fn get_all_players_in_league(conn: Data<Arc<MySqlPool>>, client: Data<Arc<Client>>, join_req: Json<JoinRequest>) -> TypedHttpResponse<Vec<Player>> {
     league_player::get_all_players_in_league(&conn, &client, join_req.0).await
 }
